Support include files in definitions
For dynamic data, it can be desirable to load external files into
variables - however, the "include" statement must reside on a separate
line. Handle this by introducing a special "include" definition, which,
if it is the only key, will be translated to a correctly formatted
"include" statement.
This patch additionally introduces empty lines between indivdidual
definition blocks.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>